completed todo

+++ src/todos/todos.service.ts
@@ -11,18 +11,20 @@ export class TodosService {
     return this.prisma.todo.create({ data: { ...createTodoDto, user_id: 1 } });
   }
 
-  async findAll() {
-    return this.prisma.todo.findMany();
+  async findAll(user_id: number) {
+    return this.prisma.todo.findMany({
+      where: { user_id },
+    });
   }
 
-  async findOne(id: number) {
-    return this.getTodo(id);
+  async findOne(id: number, user_id: number) {
+    return this.getTodo(id, user_id);
   }
 
-  async update(id: number, updateTodoDto: UpdateTodoDto) {
-    await this.getTodo(id);
+  async update(id: number, updateTodoDto: UpdateTodoDto, user_id: number) {
+    await this.getTodo(id, user_id);
     return this.prisma.todo.update({
-      where: { id },
+      where: { id, user_id },
       data: {
         title: updateTodoDto.title,
         ...(updateTodoDto.description && {
@@ -33,13 +35,13 @@ export class TodosService {
     });
   }
 
-  async remove(id: number) {
-    await this.getTodo(id);
-    return this.prisma.todo.delete({ where: { id } });
+  async remove(id: number, user_id: number) {
+    await this.getTodo(id, user_id);
+    return this.prisma.todo.delete({ where: { id, user_id } });
   }
 
-  private async getTodo(id: number) {
-    const todo = await this.prisma.todo.findFirst({ where: { id } });
+  private async getTodo(id: number, user_id: number) {
+    const todo = await this.prisma.todo.findFirst({ where: { id, user_id } });
 
     if (!todo) {
       throw new NotFoundException('Todo not found');
